requirements.txt
telethon>=1.34.0
mysql-connector-python>=8.0.33
